requirements.txt
mss>=9.0.1
pynput>=1.7.6
psutil>=5.9.6
Pillow>=10.0.0

